FE code refactoring
