+++ frontend/app/src/main/java/com/cpen321/usermanagement/ui/components/student/ReturnJobAddressComponents.kt
@@ -9,6 +9,7 @@ import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import com.cpen321.usermanagement.ui.components.common.AddressAutocompleteField
+import com.cpen321.usermanagement.ui.components.common.SelectedAddress
 
 data class AddressSelectionActions(
     val onUseCustomAddressChange: (Boolean) -> Unit,
